Move `ViewModel`s to separated package
