+++ app/src/main/java/org/sedo/satmesh/ui/ChatFragment.java
@@ -35,6 +35,8 @@ import org.sedo.satmesh.model.Message;
 import org.sedo.satmesh.model.Node;
 import org.sedo.satmesh.nearby.NearbySignalMessenger;
 import org.sedo.satmesh.ui.adapter.ChatAdapter;
+import org.sedo.satmesh.ui.vm.ChatViewModel;
+import org.sedo.satmesh.ui.vm.ViewModelFactory;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -65,7 +67,7 @@ public class ChatFragment extends Fragment {
 			View decorView = requireActivity().getWindow().getDecorView();
 			View actionModeView = decorView.findViewById(androidx.appcompat.R.id.action_mode_bar);
 			if (actionModeView != null) {
-				actionModeView.setBackgroundColor(ContextCompat.getColor(requireContext(), R.color.secondary));
+				actionModeView.setBackgroundColor(ContextCompat.getColor(requireContext(), R.color.toolbar_background));
 			}
 			return true;
 		}
@@ -317,7 +319,7 @@ public class ChatFragment extends Fragment {
 			}
 		});
 		if (toolbar.getOverflowIcon() != null) {
-			toolbar.getOverflowIcon().setTint(ContextCompat.getColor(requireContext(), R.color.on_secondary));
+			toolbar.getOverflowIcon().setTint(ContextCompat.getColor(requireContext(), R.color.toolbar_text));
 		}
 
 		toolbar.setOnMenuItemClickListener(menuItem -> {
@@ -466,7 +468,7 @@ public class ChatFragment extends Fragment {
 			}
 			if (Boolean.TRUE.equals(isActive)) {
 				binding.chatToolbar.setSubtitle(R.string.status_secure_session_active);
-				binding.chatToolbar.setSubtitleTextColor(ContextCompat.getColor(requireContext(), R.color.on_secondary));
+				binding.chatToolbar.setSubtitleTextColor(ContextCompat.getColor(requireContext(), R.color.toolbar_text));
 			}
 		});
