+++ app/src/main/java/org/sedo/satmesh/ui/vm/ChatListViewModel.java
@@ -1,4 +1,4 @@
-package org.sedo.satmesh.ui;
+package org.sedo.satmesh.ui.vm;
 
 import android.app.Application;
 import android.util.Log;
